Setup pool and test for Level2 backgrounds
WIP #735
