+++ jwst/associations/tests/data/pool_009_spec_miri_lrs_fsbkg.csv
@@ -1,0 +1,14 @@
+# Pool.Spectrum.MIRI.LRS.Background
+#
+# Examples of MIRI LRS Fixed Slit with Background
+#
+# OBS_NUM and AC_CANDIDATE must be computed and inserted. Other columns which also incorporate these values are not modified.
+#
+FILENAME|OBS_ID|PROGRAM|OBS_NUM|VISIT|VISIT_ID|VISITGRP|WFSVISIT|SEQ_ID|ACT_ID|EXPOSURE|EXP_TYPE|NEXPOSUR|EXPCOUNT|INSTRUME|DETECTOR|CHANNEL|TARGETID|TARGPROP|TARGNAME|TARGTYPE|TEMPLATE|PNTGTYPE|PNTG_SEQ|TARGORDN|EXPSPCIN|DITHPTIN|MOSTILNO|MODULE|FILTER|PUPIL|DITHERID|PATTTYPE|PATTSTRT|NUMDTHPT|PATTSIZE|SUBPXPNS|PATT_NUM|SUBPXNUM|SUBPIXEL|APERNAME|SDP_VER|SUBARRAY|GRATING|FXD_SLIT|BAND|ASN_CANDIDATE|BACKGROUND
+#
+# ########
+# Proposal 80500, 80600
+#
+# Mode: LRS Fixed Slit with Background
+@!fmt_fname(expnum)|V99009013001P0000000002101|99009|@!next(obsnum)|1|99009013001|2|NULL|1|1|1|MIR_LRS-FIXEDSLIT|NULL|NULL|MIRI|MIRIMAGE|NULL|1|FRED-AB-DOR-C|NULL|FIXED|MIRI Low Resolution Spectroscopy|SCIENCE|1|1|1|1|1|NULL|P750L|NULL|NULL|Point Source|NULL|NULL|NULL|NULL|NULL|NULL|NULL|MIRIM_FULL_SLITCNTR|2015_1|FULL|NULL|NULL|NULL|@!fmt_cand([(obsnum.value, 'OBSERVATION'), ('c1001', 'BACKGROUND')])|NULL
+@!fmt_fname(expnum)|V99009014001P0000000002101|99009|@!next(obsnum)|1|99009014001|2|NULL|1|1|1|MIR_LRS-FIXEDSLIT|NULL|NULL|MIRI|MIRIMAGE|NULL|1|FRED-AB-DOR-C|NULL|FIXED|MIRI Low Resolution Spectroscopy|SCIENCE|1|1|1|1|1|NULL|P750L|NULL|NULL|Point Source|NULL|NULL|NULL|NULL|NULL|NULL|NULL|MIRIM_FULL_SLITCNTR|2015_1|FULL|NULL|NULL|NULL|@!fmt_cand([(obsnum.value, 'OBSERVATION'), ('c1001', 'BACKGROUND')])|BACKGROUND
